Make the handlers generic.
