+++ packages/editor/src/lib/editor/types/external-content.ts
@@ -1,4 +1,3 @@
-import { TLEmbedDefinition } from '@tldraw/tlschema'
 import { VecLike } from '../../primitives/Vec'
 import { TLContent } from './clipboard-types'
 
@@ -24,7 +23,7 @@ export type TLExternalContentSource =
 	  }
 
 /** @public */
-export type TLExternalContent = {
+export type TLExternalContent<EmbedDefinition> = {
 	sources?: TLExternalContentSource[]
 	point?: VecLike
 } & (
@@ -48,7 +47,7 @@ export type TLExternalContent = {
 	| {
 			type: 'embed'
 			url: string
-			embed: TLEmbedDefinition
+			embed: EmbedDefinition
 	  }
 )
