+++ packages/tldraw/src/index.ts
@@ -26,7 +26,12 @@ export { TldrawShapeIndicators } from './lib/canvas/TldrawShapeIndicators'
 export { defaultBindingUtils } from './lib/defaultBindingUtils'
 export {
 	DEFAULT_EMBED_DEFINITIONS,
+	embedShapePermissionDefaults,
+	type CustomEmbedDefinition,
 	type DefaultEmbedDefinitionType,
+	type EmbedDefinition,
+	type TLEmbedDefinition,
+	type TLEmbedShapePermissions,
 } from './lib/defaultEmbedDefinitions'
 export {
 	centerSelectionAroundPoint,
